Create README.md
Today I spent a few hours setting up Kali Linux on a Raspberry Pi 3. The end goal is to be able to have it accessible via SSH login with root access and work headlessly. I was able to get Kali to work on the device and grant SSH access but ran into a small wall when trying to log into Kali without a password. Tomorrow is a new day and I'm confident I will be able to move forward.